# docs/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/preferences": {
      "get": {
        "operationId": "PreferencesController_get_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Preferences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "theme": {
                          "type": "string",
                          "enum": [
                            "dark",
                            "light",
                            "system"
                          ]
                        },
                        "density": {
                          "type": "string",
                          "enum": [
                            "compact",
                            "comfortable"
                          ]
                        },
                        "optionChain": {
                          "type": "object",
                          "properties": {
                            "strikeWindow": {
                              "anyOf": [
                                {
                                  "type": "number",
                                  "enum": [
                                    10
                                  ]
                                },
                                {
                                  "type": "number",
                                  "enum": [
                                    20
                                  ]
                                },
                                {
                                  "type": "number",
                                  "enum": [
                                    30
                                  ]
                                },
                                {
                                  "type": "string",
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            },
                            "columnPreset": {
                              "type": "string",
                              "enum": [
                                "core",
                                "extended",
                                "greeks"
                              ]
                            }
                          },
                          "required": [
                            "strikeWindow",
                            "columnPreset"
                          ],
                          "additionalProperties": false
                        },
                        "chart": {
                          "type": "object",
                          "properties": {
                            "interval": {
                              "type": "string",
                              "enum": [
                                "1m",
                                "5m",
                                "15m",
                                "1h",
                                "1d"
                              ]
                            },
                            "indicators": {
                              "maxItems": 8,
                              "type": "array",
                              "items": {
                                "type": "string",
                                "maxLength": 16
                              }
                            }
                          },
                          "required": [
                            "interval",
                            "indicators"
                          ],
                          "additionalProperties": false
                        },
                        "notifications": {
                          "type": "object",
                          "properties": {
                            "browser": {
                              "type": "boolean"
                            },
                            "sound": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "browser",
                            "sound"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "Session preferences",
        "tags": [
          "preferences"
        ]
      },
      "put": {
        "operationId": "PreferencesController_put_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "theme": {
                    "type": "string",
                    "enum": [
                      "dark",
                      "light",
                      "system"
                    ]
                  },
                  "density": {
                    "type": "string",
                    "enum": [
                      "compact",
                      "comfortable"
                    ]
                  },
                  "optionChain": {
                    "type": "object",
                    "properties": {
                      "strikeWindow": {
                        "anyOf": [
                          {
                            "type": "number",
                            "enum": [
                              10
                            ]
                          },
                          {
                            "type": "number",
                            "enum": [
                              20
                            ]
                          },
                          {
                            "type": "number",
                            "enum": [
                              30
                            ]
                          },
                          {
                            "type": "string",
                            "nullable": true,
                            "enum": [
                              null
                            ]
                          }
                        ]
                      },
                      "columnPreset": {
                        "type": "string",
                        "enum": [
                          "core",
                          "extended",
                          "greeks"
                        ]
                      }
                    },
                    "required": [
                      "strikeWindow",
                      "columnPreset"
                    ],
                    "additionalProperties": false
                  },
                  "chart": {
                    "type": "object",
                    "properties": {
                      "interval": {
                        "type": "string",
                        "enum": [
                          "1m",
                          "5m",
                          "15m",
                          "1h",
                          "1d"
                        ]
                      },
                      "indicators": {
                        "maxItems": 8,
                        "type": "array",
                        "items": {
                          "type": "string",
                          "maxLength": 16
                        }
                      }
                    },
                    "required": [
                      "interval",
                      "indicators"
                    ],
                    "additionalProperties": false
                  },
                  "notifications": {
                    "type": "object",
                    "properties": {
                      "browser": {
                        "type": "boolean"
                      },
                      "sound": {
                        "type": "boolean"
                      }
                    },
                    "required": [
                      "browser",
                      "sound"
                    ],
                    "additionalProperties": false
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Merged preferences",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "theme": {
                          "type": "string",
                          "enum": [
                            "dark",
                            "light",
                            "system"
                          ]
                        },
                        "density": {
                          "type": "string",
                          "enum": [
                            "compact",
                            "comfortable"
                          ]
                        },
                        "optionChain": {
                          "type": "object",
                          "properties": {
                            "strikeWindow": {
                              "anyOf": [
                                {
                                  "type": "number",
                                  "enum": [
                                    10
                                  ]
                                },
                                {
                                  "type": "number",
                                  "enum": [
                                    20
                                  ]
                                },
                                {
                                  "type": "number",
                                  "enum": [
                                    30
                                  ]
                                },
                                {
                                  "type": "string",
                                  "nullable": true,
                                  "enum": [
                                    null
                                  ]
                                }
                              ]
                            },
                            "columnPreset": {
                              "type": "string",
                              "enum": [
                                "core",
                                "extended",
                                "greeks"
                              ]
                            }
                          },
                          "required": [
                            "strikeWindow",
                            "columnPreset"
                          ],
                          "additionalProperties": false
                        },
                        "chart": {
                          "type": "object",
                          "properties": {
                            "interval": {
                              "type": "string",
                              "enum": [
                                "1m",
                                "5m",
                                "15m",
                                "1h",
                                "1d"
                              ]
                            },
                            "indicators": {
                              "maxItems": 8,
                              "type": "array",
                              "items": {
                                "type": "string",
                                "maxLength": 16
                              }
                            }
                          },
                          "required": [
                            "interval",
                            "indicators"
                          ],
                          "additionalProperties": false
                        },
                        "notifications": {
                          "type": "object",
                          "properties": {
                            "browser": {
                              "type": "boolean"
                            },
                            "sound": {
                              "type": "boolean"
                            }
                          },
                          "required": [
                            "browser",
                            "sound"
                          ],
                          "additionalProperties": false
                        }
                      },
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "data"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "400": {
            "description": "Validation failed (`VALIDATION_FAILED`, per-field issues in `details.issues`)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid credentials",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "description": "Stable machine-readable code, e.g. VALIDATION_FAILED"
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {
                          "type": "object",
                          "additionalProperties": {}
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ],
                      "additionalProperties": false
                    }
                  },
                  "required": [
                    "error"
                  ],
                  "additionalProperties": false
                }
              }
            }
          }
        },
        "summary": "Merge preferences (unknown keys rejected)",
        "tags": [
          "preferences"
        ]
      }
    },
...
  },
  "info": {
    "title": "TradeOS API",
    "description": "Indian market option-chain terminal — REST contract v1.\n\nEvery response is JSON. Errors share one envelope: `{ \"error\": { \"code\", \"message\", \"details?\" } }`.\n\nRealtime market data does not use REST: see the WebSocket protocol in `docs/websocket-protocol.md` (`/ws/market`).\n\nInstrument keys are TradeOS-owned (`NSE:INDEX:NIFTY50`, `NSE:OPT:NIFTY50:2026-09-15:24000:CE`); provider symbols never appear in this API.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "operator-key": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Operator-Key"
      }
    },
...
  }
}